Remove the coroutine launch from LoadPostsUseCase
